User commit

+++ src/main/java/com/example/demo/entity/StudentEntity.java
@@ -3,8 +3,11 @@ import jakarta.persistence.Entity;
 
 @Entity
 public class StudenEntity{
-    private int id;
+    
+    private Long id;
     private String name;
     private String email;
     private float cgpa;
+
+
 }
